save filters with name

+++ lib/2_application/options/library/option_int_ranges.dart
@@ -14,7 +14,7 @@ enum OptionIntRange {
   });
 
 
-  final double low;
-  final double high;
+  final int low;
+  final int high;
   final int steps;
 }
